requirements.txt
python-dotenv
psycopg2-binary
pyyaml
pyquery
aiohttp
